medi_agenda/medi_agenda/requirements.txt
flask
mysql-connector-python
google-generativeai
fpdf
python-dotenv
